Added back helpers.js scroll method
Added back helpers.js scroll method and fixed its call executions so it didn't require elementHandles.  Tests pass and scrolling works.
