+++ tests/milo/failedblock.test.js
@@ -4,32 +4,10 @@ const { expect, test } = require('@playwright/test');
 const parse = require('../../libs/parse.js');
 const failedBlock = require('../../features/milo/failedblock.spec.js');
 const selectors = require('../../selectors/milo/failedblock.selectors.js');
+const helpers = require('../../libs/helpers.js');
 
 // Parse the feature file into something flat that can be tested separately
 const { name, features } = parse(failedBlock);
-
-/**
- * Slow/fast scroll JS evaluation method.
- * To be used in page.evaluate, i.e. page.evaluate(scroll, { direction: 'value', speed: 'value' });
- * @param direction string direction you want to scroll
- * @param speed string speed you would like to scroll. Options: slow, fast
-*/
-const scroll = async (args) => {
-  const { direction, speed } = args;
-  // eslint-disable-next-line no-promise-executor-return
-  const delay = (ms) => new Promise((resolve) => setTimeout(resolve, ms));
-  const scrollHeight = () => document.body.scrollHeight;
-  const start = direction === 'down' ? 0 : scrollHeight();
-  const shouldStop = (position) => (direction === 'down' ? position > scrollHeight() : position < 0);
-  const increment = direction === 'down' ? 100 : -100;
-  const delayTime = speed === 'slow' ? 30 : 5;
-  console.error(start, shouldStop(start), increment);
-  for (let i = start; !shouldStop(i); i += increment) {
-    window.scrollTo(0, i);
-    // eslint-disable-next-line no-await-in-loop
-    await delay(delayTime);
-  }
-};
 
 test.describe(`${name}`, () => {
   features.forEach((props) => {
@@ -40,8 +18,8 @@ test.describe(`${name}`, () => {
 
       // Added scrolling for failed block JS to load.
       // Without it, test provides false count for validation checking.
-      await page.evaluate(scroll, { direction: 'down', speed: 'slow' });
-      await page.evaluate(scroll, { direction: 'up', speed: 'fast' });
+      await page.evaluate(helpers.scroll, { direction: 'down', speed: 'slow' });
+      await page.evaluate(helpers.scroll, { direction: 'up', speed: 'fast' });
 
       const locator = page.locator(selectors[props.tag]);
       const count = await locator.count();
